Compile of more warning fixes under the latest-stable version of gcc 4.5.2.
As a part of this change I removed LONG_LONG from ArgParser as it was not really
used and was potential for some compatibility issue under MSVC 8.0+ and pgCC.

+++ src/clustering/TransMSTClusterMaker.cpp
@@ -83,7 +83,7 @@ TransMSTClusterMaker::initialize() {
     
     // Reserve space for TransCacheEntrys
     ASSERT ( estList != NULL );
-    const int EstListSize = estList->size();
+    const size_t EstListSize = estList->size();
     metricCache.reserve(EstListSize);
     metricCache.insert(metricCache.begin(), EstListSize, NULL);
     // Everything went well.
